@@ -1,26 +1,84 @@
 {
+  "//": "Generated by projen. To modify, edit .projenrc.js and run \"npx projen\".",
   "name": "cdk8s",
-  "version": "0.0.0",
   "description": "Cloud Development Kit for Kubernetes",
-  "scripts": {
-    "build": "jsii && npm run docgen",
-    "watch": "jsii -w",
-    "package": "jsii-pacmak",
-    "lint": "eslint . --ext .ts",
-    "test": "yarn lint && jest",
-    "docgen": "jsii-docgen"
+  "main": "lib/index.js",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/awslabs/cdk8s.git"
   },
+  "bin": {},
+  "scripts": {
+    "projen": "node .projenrc.js && yarn install",
+    "projen:upgrade": "yarn upgrade projen && yarn projen",
+    "test": "yarn eslint && rm -fr lib/ && jest --passWithNoTests && yarn compile",
+    "compile": "jsii --silence-warnings=reserved-word && jsii-docgen && cp src/_loadurl.js lib/",
+    "watch": "jsii -w --silence-warnings=reserved-word",
+    "compat": "npx jsii-diff npm:$(node -p \"require('./package.json').name\")",
+    "package": "jsii-pacmak",
+    "build": "yarn compile",
+    "eslint": "eslint . --ext .ts",
+    "docgen": "jsii-docgen",
+    "test:watch": "jest --watch",
+    "test:update": "jest -u"
+  },
+  "author": {
+    "name": "Amazon Web Services",
+    "url": "https://aws.amazon.com"
+  },
+  "devDependencies": {
+    "constructs": "2.0.1",
+    "@types/follow-redirects": "^1.8.0",
+    "@types/json-stable-stringify": "^1.0.32",
+    "@types/yaml": "^1.2.0",
+    "json-schema-to-typescript": "^8.0.1",
+    "projen": "^0.2.3",
+    "standard-version": "^8.0.0",
+    "jsii": "^1.6.0",
+    "jsii-diff": "^1.6.0",
+    "jsii-pacmak": "^1.6.0",
+    "@types/node": "^10.17.0",
+    "typescript": "^3.8.3",
+    "@typescript-eslint/eslint-plugin": "^2.31.0",
+    "@typescript-eslint/parser": "^2.19.2",
+    "eslint": "^6.8.0",
+    "eslint-import-resolver-node": "^0.3.3",
+    "eslint-import-resolver-typescript": "^2.0.0",
+    "eslint-plugin-import": "^2.20.2",
+    "json-schema": "^0.2.5",
+    "jsii-docgen": "^1.3.2",
+    "jest": "^26.0.1",
+    "@types/jest": "^25.2.1",
+    "ts-jest": "^26.1.0"
+  },
+  "peerDependencies": {
+    "constructs": "2.0.1"
+  },
+  "dependencies": {
+    "follow-redirects": "^1.11.0",
+    "json-stable-stringify": "^1.0.1",
+    "yaml": "^1.7.2"
+  },
+  "bundledDependencies": [
+    "yaml",
+    "json-stable-stringify",
+    "follow-redirects"
+  ],
   "keywords": [
     "cdk",
     "kubernetes",
     "k8s",
     "constructs"
   ],
-  "main": "lib/index.js",
+  "engines": {
+    "node": ">= 10.17.0"
+  },
+  "license": "Apache-2.0",
+  "version": "0.0.0",
   "types": "lib/index.d.ts",
+  "stability": "experimental",
   "jsii": {
     "outdir": "dist",
-    "versionFormat": "short",
     "targets": {
       "java": {
         "package": "org.cdk8s",
