@@ -1,5 +1,5 @@
 import { Construct, Node, IConstruct } from 'constructs';
-import fs = require('fs');
+import * as fs from 'fs';
 import { Chart } from './chart';
 import * as path from 'path';
 import { Yaml } from './yaml';
@@ -107,12 +107,12 @@ export class App extends Construct {
 
 function validate(app: App) {
 
-    // Note this is a copy-paste of https://github.com/aws/constructs/blob/master/lib/construct.ts#L438.
-    const errors = Node.of(app).validate();
-    if (errors.length > 0) {
-      const errorList = errors.map(e => `[${Node.of(e.source).path}] ${e.message}`).join('\n  ');
-      throw new Error(`Validation failed with the following errors:\n  ${errorList}`);
-    }
+  // Note this is a copy-paste of https://github.com/aws/constructs/blob/master/lib/construct.ts#L438.
+  const errors = Node.of(app).validate();
+  if (errors.length > 0) {
+    const errorList = errors.map(e => `[${Node.of(e.source).path}] ${e.message}`).join('\n  ');
+    throw new Error(`Validation failed with the following errors:\n  ${errorList}`);
+  }
 
 }
 
