@@ -1,10 +1,10 @@
-import { Tokenization, DefaultTokenResolver, StringConcat, Construct } from "constructs";
+import { Tokenization, DefaultTokenResolver, StringConcat, Construct } from 'constructs';
 
 const TOKEN_RESOLVER = new DefaultTokenResolver(new StringConcat());
 
 export function resolve<T>(scope: Construct, obj: T): T {
   return Tokenization.resolve(obj, {
     scope,
-    resolver: TOKEN_RESOLVER
+    resolver: TOKEN_RESOLVER,
   });
 }
